pyproject.toml
[project]
name = "stock-data-skill"
version = "0.0.1"
description = "CLI and OpenClaw skill for stock/crypto data analysis"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["stock", "crypto", "analysis", "openclaw", "cli", "akshare"]
dependencies = [
    "open-stock-data>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
stock-data = "stock_data_skill.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stock_data_skill"]
